Remove cglib

+++ cpf-api-app/src/main/java/ca/bc/gov/open/cpf/api/web/builder/BatchJobUiBuilder.java
@@ -8,10 +8,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -22,7 +19,6 @@ import ca.bc.gov.open.cpf.api.domain.BatchJob;
 import ca.bc.gov.open.cpf.api.domain.BatchJobExecutionGroup;
 import ca.bc.gov.open.cpf.api.domain.BatchJobResult;
 import ca.bc.gov.open.cpf.api.scheduler.BatchJobService;
-import ca.bc.gov.open.cpf.api.security.CpfMethodSecurityExpressions;
 import ca.bc.gov.open.cpf.plugin.impl.BusinessApplication;
 import ca.bc.gov.open.cpf.plugin.impl.module.Module;
 
@@ -37,8 +33,7 @@ import com.revolsys.util.DateUtil;
 import com.revolsys.util.JavaBeanUtil;
 
 @Controller
-public class BatchJobUiBuilder extends CpfUiBuilder implements
-  CpfMethodSecurityExpressions {
+public class BatchJobUiBuilder extends CpfUiBuilder {
 
   public BatchJobUiBuilder() {
     super("batchJob", BatchJob.BATCH_JOB, BatchJob.BATCH_JOB_ID, "Batch Job",
@@ -88,7 +83,7 @@ public class BatchJobUiBuilder extends CpfUiBuilder implements
 
     final BusinessApplication businessApplication = getBusinessApplication(businessApplicationName);
     final Module module = businessApplication.getModule();
-    final ModuleUiBuilder appBuilder = getBuilder(Module.class);
+    final CpfUiBuilder appBuilder = getBuilder(Module.class);
 
     final Map<String, String> parameterKeys = new HashMap<String, String>();
     parameterKeys.put("moduleName", "name");
@@ -99,10 +94,9 @@ public class BatchJobUiBuilder extends CpfUiBuilder implements
     "/admin/jobs"
   }, method = RequestMethod.GET)
   @ResponseBody
-  @PreAuthorize(ADMIN_OR_MODULE_ANY_ADMIN_EXCEPT_SECURITY)
-  @Transactional(propagation = Propagation.REQUIRES_NEW, readOnly = true)
   public Object pageList(final HttpServletRequest request,
     final HttpServletResponse response) throws IOException {
+    checkAdminOrAnyModuleAdminExceptSecurity();
     HttpServletUtils.setAttribute("title", "Batch Jobs");
     return createDataTableHandler(request, "list");
   }
@@ -111,12 +105,11 @@ public class BatchJobUiBuilder extends CpfUiBuilder implements
     "/admin/modules/{moduleName}/apps/{businessApplicationName}/jobs"
   }, method = RequestMethod.GET)
   @ResponseBody
-  @PreAuthorize(ADMIN_OR_ADMIN_FOR_MODULE)
-  @Transactional(propagation = Propagation.REQUIRES_NEW, readOnly = true)
   public Object pageModuleAppList(final HttpServletRequest request,
     final HttpServletResponse response, @PathVariable final String moduleName,
     @PathVariable final String businessApplicationName) throws IOException,
     ServletException {
+    checkAdminOrModuleAdmin(moduleName);
     getModuleBusinessApplication(moduleName, businessApplicationName);
 
     final Map<String, Object> parameters = new HashMap<String, Object>();
@@ -133,13 +126,12 @@ public class BatchJobUiBuilder extends CpfUiBuilder implements
     "/admin/modules/{moduleName}/apps/{businessApplicationName}/jobs/{batchJobId}"
   }, method = RequestMethod.GET)
   @ResponseBody
-  @PreAuthorize(ADMIN_OR_MODULE_ADMIN)
-  @Transactional(propagation = Propagation.REQUIRES_NEW, readOnly = true)
   public ElementContainer pageModuleAppView(final HttpServletRequest request,
     final HttpServletResponse response, @PathVariable final String moduleName,
     @PathVariable final String businessApplicationName,
     @PathVariable final Integer batchJobId) throws IOException,
     ServletException {
+    checkAdminOrAnyModuleAdmin();
     getModuleBusinessApplication(moduleName, businessApplicationName);
     final DataObject batchJob = getBatchJob(businessApplicationName, batchJobId);
 
@@ -163,8 +155,8 @@ public class BatchJobUiBuilder extends CpfUiBuilder implements
   }, method = RequestMethod.POST)
   public void postClientCancel(@PathVariable final long batchJobId) {
     final String consumerKey = getConsumerKey();
-    final DataObject batchJob = getCpfDataAccessObject().getBatchJob(
-      consumerKey, batchJobId);
+    final DataObject batchJob = getDataAccessObject().getBatchJob(consumerKey,
+      batchJobId);
     if (batchJob == null) {
       throw new PageNotFoundException("The cloud job " + batchJobId
         + " does not exist");
@@ -180,8 +172,8 @@ public class BatchJobUiBuilder extends CpfUiBuilder implements
   }, method = RequestMethod.POST)
   public void postClientDelete(@PathVariable final long batchJobId) {
     final String consumerKey = getConsumerKey();
-    final DataObject batchJob = getCpfDataAccessObject().getBatchJob(
-      consumerKey, batchJobId);
+    final DataObject batchJob = getDataAccessObject().getBatchJob(consumerKey,
+      batchJobId);
     if (batchJob == null) {
       throw new PageNotFoundException("The cloud job " + batchJobId
         + " does not exist");
@@ -196,12 +188,11 @@ public class BatchJobUiBuilder extends CpfUiBuilder implements
       value = {
         "/admin/modules/{moduleName}/apps/{businessApplicationName}/jobs/{batchJobId}/cancel"
       }, method = RequestMethod.POST)
-  @PreAuthorize(ADMIN_OR_MODULE_ADMIN)
-  @Transactional(propagation = Propagation.REQUIRES_NEW)
   public void postModuleAppCancel(final HttpServletRequest request,
     final HttpServletResponse response, @PathVariable final String moduleName,
     @PathVariable final String businessApplicationName,
     @PathVariable final Long batchJobId) throws IOException, ServletException {
+    checkAdminOrAnyModuleAdmin();
     getModuleBusinessApplication(moduleName, businessApplicationName);
     final BatchJobService batchJobService = getBatchJobService();
     batchJobService.cancelBatchJob(batchJobId);
@@ -217,12 +208,11 @@ public class BatchJobUiBuilder extends CpfUiBuilder implements
       value = {
         "/admin/modules/{moduleName}/apps/{businessApplicationName}/jobs/{batchJobId}/delete"
       }, method = RequestMethod.POST)
-  @PreAuthorize(ADMIN_OR_MODULE_ADMIN)
-  @Transactional(propagation = Propagation.REQUIRES_NEW)
   public void postModuleAppDelete(final HttpServletRequest request,
     final HttpServletResponse response, @PathVariable final String moduleName,
     @PathVariable final String businessApplicationName,
     @PathVariable final Long batchJobId) throws IOException, ServletException {
+    checkAdminOrAnyModuleAdmin();
     getModuleBusinessApplication(moduleName, businessApplicationName);
     final BatchJobService batchJobService = getBatchJobService();
     batchJobService.deleteBatchJob(batchJobId);
